downgrade app engine plugin
This downgrade is necessary since we are using the extraction of the fat jar within the staging directory and the original jar needs to be removed using the workaround by adding the <artifact>${project.build.directory}</artifact> which points to the target directory. JHipster generator has moved to a different approach of using thin jars for deployment which we need to adapt in the future.